feature-BJS2-26145: fix + навела красоты
