+++ src/main/java/faang/school/postservice/mapper/comment/CommentMapper.java
@@ -1,4 +1,4 @@
-package faang.school.postservice.mapper;
+package faang.school.postservice.mapper.comment;
 
 
 import faang.school.postservice.dto.comment.CommentDto;
