+++ src/main/java/faang/school/postservice/model/Post.java
@@ -1,7 +1,7 @@
 package faang.school.postservice.model;
 
 import faang.school.postservice.model.ad.Ad;
-import faang.school.postservice.moderation.Verifyible;
+import faang.school.postservice.moderation.Verifiable;
 import jakarta.persistence.*;
 
 import lombok.AllArgsConstructor;
@@ -22,7 +22,7 @@ import java.util.List;
 @Builder
 @Entity
 @Table(name = "post")
-public class Post implements Verifyible {
+public class Post implements Verifiable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -82,19 +82,4 @@ public class Post implements Verifyible {
     @Temporal(TemporalType.TIMESTAMP)
     @Column(name = "verified_date")
     private LocalDateTime verifiedDate;
-
-    @Override
-    public void setVerificationValue(boolean result) {
-        this.setVerified(result);
-    }
-
-    @Override
-    public void initVerifiedDate() {
-        this.setVerifiedDate(LocalDateTime.now());
-    }
-
-    @Override
-    public String getContentText() {
-        return this.getContent();
-    }
 }
