+++ src/main/java/faang/school/postservice/moderation/Verifiable.java
@@ -1,0 +1,12 @@
+package faang.school.postservice.moderation;
+
+import java.time.LocalDateTime;
+
+public interface Verifiable {
+
+    void setVerified(Boolean verified);
+
+    void setVerifiedDate(LocalDateTime verifiedDate);
+
+    String getContent();
+}
