--- src/main/java/faang/school/postservice/moderation/Verifyible.java
@@ -1,10 +1,0 @@
-package faang.school.postservice.moderation;
-
-public interface Verifyible {
-
-    void setVerificationValue(boolean result);
-
-    void initVerifiedDate();
-
-    String getContentText();
-}
