+++ src/main/java/faang/school/postservice/service/post/PostService.java
@@ -15,7 +15,6 @@ import java.util.List;
 import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutorService;
-import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
